requirements.txt
opencv-python-headless
numpy
Flask
Pillow
